requirements.txt
telethon
pymongo
python-dotenv
fastapi
uvicorn[standard]
tweepy
python-dotenv
minio
httpx
